requirements.txt
numpy==1.20.1
Pillow==8.1.0
tqdm==4.56.2
